Update Fusion command implementation and version to 1.1.10
