+++ src/main/resources/paper-plugin.yml
@@ -1,5 +1,5 @@
 name: YVtils-SMP
-version: 1.1.8
+version: 1.1.10
 main: yv.tils.smp.YVtils
 api-version: '1.21'
 author: WolfiiYV
